+++ lib/trailblazer/activity/task_wrap/extension.rb
@@ -41,7 +41,7 @@ module Trailblazer
 
         # Merges {extension_rows} into the {Pipeline} instance.
         # This is usually used in step extensions or at runtime for {wrap_runtime}.
-        def call(task_wrap_pipeline)
+        def call(task_wrap_pipeline, **)
           Adds.apply_adds(task_wrap_pipeline, @extension_rows)
         end
 
@@ -49,35 +49,11 @@ module Trailblazer
         # of a step in an activity. The returned extensionn can be passed directly via {:extensions}
         # to the compiler, or when using the `#step` DSL.
         def self.WrapStatic(*inserts)
-          WrapStatic.new(extension: TaskWrap.Extension(*inserts))
+          Activity::Deprecate.warn caller_locations[0], "Using `TaskWrap::Extension.WrapStatic()` is deprecated. Please use `TaskWrap.Extension()`."
+
+          # FIXME: change or deprecate.
+          TaskWrap.Extension(*inserts)
         end
-
-        # Extension are used at compile-time with {wrap_static}, mostly with the {dsl} gem.
-        # At compile-time, {WrapStatic} extensions are compiled in {Sequence::Compiler}.
-        #
-        # Each extension alters the activity's wrap_static taskWrap.
-        # {WrapStatic#call} returns the new "config" object with an updated config[:wrap_static] field.
-        class WrapStatic
-          def initialize(extension:)
-            @extension = extension
-          end
-
-          # DISCUSS: `wrap_static[task]` is too implicit, we're returning the default tw here set via Hash.new(). the initial tw should probably be generated explicitly via the DSL?
-          def call(config:, task:, **)
-            wrap_static = config[:wrap_static]  # The activity's data structure keeping a map of {task => task_wrap}. This will be
-                                                # merged into {config}
-            task_wrap   = wrap_static[task]     # the "actual" taskWrap for {task}. Your steps will be added to that.
-
-            # Execute the extension, which adds steps to the tw.
-            new_task_wrap = @extension.(task_wrap)
-
-            # Overwrite the original task_wrap:
-            wrap_static = wrap_static.merge(task => new_task_wrap)
-
-            # DISCUSS: does it make sense to test the immutable behavior here? What would be the worst outcome?
-            config.merge(wrap_static: wrap_static) # Return new config hash. This needs to be immutable code!
-          end
-        end # WrapStatic
       end # Extension
     end # TaskWrap
   end