+++ test/wrap/task_wrap_test.rb
@@ -154,6 +154,27 @@ class TaskWrapTest < Minitest::Spec
     return wrap_ctx, original_args
   end
 
+  it "deprecates {TaskWrap.WrapStatic}" do
+    adds = [
+      [method(:add_1), id: "user.add_1", prepend: "task_wrap.call_task"],
+      [method(:add_2), id: "user.add_2", append:  "task_wrap.call_task"],
+    ]
+
+    ext = nil
+    _, warning = capture_io do
+      ext = TaskWrap::Extension.WrapStatic(*adds)
+    end
+    line_number_for_binary = __LINE__ - 2
+
+    # lines = warning.split("\n")
+    # lines[0] = lines[0][0..-5]+"." if lines[0] =~ /\d-\d+-\d/
+    # warning = lines.join("\n")
+
+    assert_equal warning, %{[Trailblazer] #{File.realpath(__FILE__)}:#{line_number_for_binary} Using `TaskWrap::Extension.WrapStatic()` is deprecated. Please use `TaskWrap.Extension()`.\n}
+    assert_equal ext.class, Trailblazer::Activity::TaskWrap::Extension
+    # DISCUSS: should we test if the extension is correct?
+  end
+
   describe "{TaskWrap.container_activity_for}" do
     it "accepts {:wrap_static} option" do
       host_activity = Activity::TaskWrap.container_activity_for(Object, wrap_static: {a: 1})